discord_bot/utils/auth.py
from werkzeug.security import generate_password_hash, check_password_hash

def hash_password(password: str) -> str:
    """パスワードをハッシュ化する関数"""
    return generate_password_hash(password)

def verify_password(stored_password: str, provided_password: str) -> bool:
    """提供されたパスワードが保存されたハッシュと一致するか確認する関数"""
    return check_password_hash(stored_password, provided_password)

def authenticate_user(username, password):
    """ユーザー認証のダミー関数"""
    # 実際の認証処理をここに実装
    return username == 'admin' and password == 'password'

def register_user(username, password):
    """ユーザー登録のダミー関数"""
    # 実際の登録処理をここに実装
    return True